Reduce usage of ns_mktemp
ns_mktemp uses the deprecated old POSIX call mktemp(), which should
not be used anymore for security reasons (race between the name
creation and opening the file). This change removes all useages of
"ns_mktemp" from the NaviServer code and replaces it with calls to the
safe Tcl call "file tempfile ..." as introduced by Tcl 8.6. A
compatibility version for Tcl 8.5 is included in the code.