Increase scalability on DB operations by reducing DB pool locks
On one of our main servers, the total number of locks and busy locks
is the highest on nsdb:pool1, where a large number of these locks
origin from statistics operations. The new code caches these
statistics in the handles (which are per thread, requiring no locks)
and transfers the aggregated values on handle closes or statistics
calls.