New crypto function: ::ns_crypto::pbkdf2_hmac
This function computes a password hash using PBKDF2 (Password-Based
Key Derivation Function 2). This function is used to reduce
vulnerabilities of brute-force attacks against password hashes and is
used e.g. in SCRAM (Salted Challenge Response Authentication
Mechanism).

The hash function of SCRAM is PBKDF2 [RFC2898] with HMAC as the
pseudorandom function (PRF) and with dkLen == output length of
HMAC == output length of the digest function.

Here is an example of using pbkdf2_hmac for the hash function of
SCRAM-sha-256. RFC 7677 recommends to use 15K iterations for PBKDF2.

  ::ns_crypto::pbkdf2_hmac \
           -digest sha256 \
           -iterations 15000 \
           -secret $password \
           -salt $salt]