make sure to prodoce "Tcl-external" UTF-8 for multi-byte characters
Earlier versions assumed that tcl-internal UTF-8 is also valid UTF-8 for external
sources, which is often, but not always true. For example, "U+1F603 Smiling Face
with Open Mouth Unicode Character" is in Tcl internal representation 6 bytes and
in external representation 4 bytes. The bypass of the conversion for utf-8
was removed by this change. Regression test was extended.