set default for "concurrentinterpcreate" to "true" for Tcl 8.6 or newer
Versions up to at least Tcl 8.5 are known to
crash in case two threads create interpreters at the same
time. These crashes were hard to reproduce, but serializing
helped. Since all our major servers are running since several
years without problems with this parameter turned on the
the default is now set to "true", when NaviServer is
compiled with Tcl 8.6 or newer.