Raise an exception when trying to use "ns_urldecode" to produce invalid UTF-8
Background: several (external) functions expect valid UTF-8 to be passed in
and crash if this is not the case. One such example is tDOM. These nasty
byte sequences are used more intensively by vulnerability scanners.

Therefore, ns_urldecode raises now an exception, when it tries to convert
to invalid UTF-8. It is still possible to convert to other charsets.

    ns_urldecode -charset iso8859-1 -part path "/mot%C3or"

When urldecode() is called internally and would produce invalid UTF-8,
it truncates the string (and writes a warning to the system log).