ns_parsequery: added option "-charset" and raise exception on failure
The new option "-charset" can be used to add a charset for the result encoding
of the passed-in HTTP query. In case the charset is UTF8 (default on
most platforms), and the content is invalid UTF-8, an exception is raised
(similar to ns_urldecode).

Extended regression test and updated manual pages.