Provide a hint when cache-entry was too large for caching
Background: the size of the entry is typically determined after
the execution of a potentially expensive query. During the
eval of the command, the cache entry is locked and forces
a serialization. However, this means that in these cases
the situation is worse than without a cache, where some
queries can be executed in parallel. We faced the situation
of an expected slowdown of the server with many
"create entry collision", where due to application matters
an entry was becoming too large. This situation is not
easy to debug, especially in stress situations. The log
message would have helped a log to identify the cause.