Added inverse operation of "ns_quotehtml" named "ns_unquotehtml"
The new command "ns_unquotehtml" replaces the named and numeric entities
in the provided string with the native values. The command is similar to
"ns_striphtml", but "ns_striphtml" removes as well other HTML markup
which might not be desired in all acses.

This change fixes as well a bug with numeric entities (the old code
assumed, these are starting directly with a number after the
ampersand) and it adds support for numeric entities with hexadecimal
values (so far with the same value range as for decimal numeric
entities).

Extended regression test.