Provide an error message when the configured locale is not installed on the host
This change causes NaviServer to abort, when the configured locale is
not installed on the host. Typically, this locale is e.g. used by
ns_strcoll for determining the default collating order. The
configuration file for the regression testing sets the environment
variable LANG to "en_US.UTF-8". This means that for running the stock
regression test, this locale must be installed on the system.