Silence warning with recent versions of gcc when certain values of _FORTIFY_SOURCE/-Wstringop-overflow are set
Newer versions of gcc support warning of dangerous operations (such as
e.g. strncat) when these depend on not easy traceable sources. In the
fixed case, the warning was:

   warning: ‘__builtin_strncat’ specified bound depends on
   the length of the source argument

With FORTIFY_SOURCE whenever possible, GCC tries to use buffer-length
aware replacements for functions, which was not possible in the case
in question. The documentation says that with _FORTIFY_SOURCE set to
2, some more checking is added, but some conforming programs might
fail.