ns_parseheader, Ns_ParseHeader(): return the field number (index) of the parsed entry
Previously, there was no easy way to determine, what has been
parsed. Now, in success cases, the function returns the index of the
entry. This function made it possible to generalize and simplify the
Tcl-level parsing of "multipart/form-data"
significantly. Additionally, a new optional argument "-prefix" was
added. When specified, it adds the specified prefix to the key.