Implementation of fallback encodings for formdata
Fallbackencodings are implemented for cases, where invalid UTF-8 is
provided in forms, although utf-8 was explicitly or implicitly
specified. This is useful for legacy or external applications
providing invalid data.

When decoding to UTF-8 fails, the list of fallbackencodings is
processed and as soon a decoding succeeds, the command returns.  This
differs from the default charset, which is used in forms, when no
charset is specified explicitly.

The new flag "-fallbackencodings" can be provided for the commands
"ns_getform" and "ns_parsequery"

  ns_getform ?-fallbackencodings encodings? ?defaultencoding?
  ns_parsequery ?-fallbackencodings encodings? ?-charset charset? /content/

So far, "-fallbackencodings" is implemented for "ns_parsequery" and
"ns_getform", where the data is provided as
"application/x-www-form-urlencoded" either when parsing from memory or
from spool file. The "multipart/form-data" implementation is not yet
covered.

extended regression test for cases testing "-fallbackencodings"