Adjust samesite behavior and support "-samesite" flag for ns_deletecookie
Mozilla and Chrome changed the default value for SameSite of cookies
from "none" to "lax" in February and Aug 2020. Cookies that explicitly
set SameSite=None in order to enable cross-site delivery must also set
the "Secure" attribute.

In order to align NaviServer, the default behavior when setting a
cookie is "lax".  When trying to set a cookie with "-samesite none",
and "-secure" was not specified, a warning is generated, and the
"-samesite lax" is assumed, since major browsers announced that these
will reject these cookies soon.

This effects as well unsetting of cookies, which follow the same
strategy.

Adjusted the regression to reflect the changes in default values.