Shorten warning message when invalid UTF-8 is found in forms.
Previously, the full buffer was written to the log file when invalid
UTF-8 was encountered. When uploading long text, these messages can be
quite long.  Now NaviServer reports a short string containing the
begin of the line, followed by the invalid sequence delimited by the
character '|', e.g.,

    Warning: decoded string contains invalid UTF-8: 'BEGIN_OF_LINE...|\xe9\xe8\xe7|...'

In this example, the line contains an invalid 3-Byte UTF-8 sequence.

In addition, the command ns_valid_utf8 got an optional 2nd argument
which contains in failure cases the shortend string with
printable binary values (like in the warning).

Extended documentation and regression test.