Support optional parameter "-fallbackcharset" for "ns_urldecode"
Similar to "ns_parsequery" and "ns_getform", the command
"ns_urldecode" accepts now the "-fallbackcharset". In opposite to the
two other commands (called implicitly), "ns_urldecode" does not user
the server configuration parameter "formFallbackCharset", since this
is not a form-specific and it has to be called expicitly in
different usage scenarios.

        ns_parsequery ?-charset charset? ?-fallbackcharset fallbackcharset? ?--? querystring
        ns_urldecode ?-charset charset? ?-fallbackcharset fallbackcharset? ?-part part? ?--? string
        ns_getform ?-fallbackcharset fallbackcharset? ?charset?

Additional changes:
- Factored out NsGetFallbackEncoding() for handling "-fallbackcharset"
- Extended documentation and regression test.