Added shrinking of database buffers
In cases, where the allocated memory was larger than a specified size, and
the actually needed amount is less than a quarter, shrink the
buffer. We do not have to use realloc(), since the content is
anyhow copied later. Note that we have to use the same
alloc()/free() functions that also Tcl uses.