New experimental command: ns_parsehtml
In opposite to HTML/XML parsers like tdom, expat etc. this function is
intended to process also invalid markup for post-processing user
input, containing potentially disallowed or incorrect (e.g., missing
closing tags) markup. The result of the command is a list of entries,
where every element contains the type of the entry ("text", "tag",
"comment"), the start and end byte position of the chunk and
optionally the content. In case of "tag" elements, the content
consists of the name of the tag and a Tcl dict containing the
attributes.

   ns_parsehtml ?-withtext? ?--? inputstring

When "-withtext" is specified, "text" and "comment" entries contains
the text, which could also be obtained from the input string. However,
at least for debugging, this option appears useful.