ns_parsehtml: next iteration
The version from a few days ago was returning byte positions of the
parsed chunk as primary instrument to get the original text
chunks. However, byte positions are not a good instrument in Tcl,
since all offsets are computed which character positions, which differ
when multibyte characters are used. So, the new version returns always
the parsed text chunks optionally with the angle brackets surrounding
tags, comments and processing instructions removed.

   ns_parsehtml ?-noangle? ?--? inputstring