Bugfix: support greater than sign ">" inside attribute values.
Previously, NaviServer determined the terminating end-of-tag character
as literally the first greater than sign, no matter if this was
used as attribute values between single or double quotes.

However, the Living Standard of HTML [1] allows the use of less than
"<" and greater than ">" signs inside attribute values as long these
are between single or double quotes.

The guide [2] just recommends using character escapes for "<", ">" and "&".

[1] https://html.spec.whatwg.org/multipage/syntax.html#syntax-attribute-value
[2] https://www.w3.org/International/questions/qa-escapes