Properly clear ADP flags in case of errors
Previously, The following ADP page could lead to a full breakdown of
the server, since the error states of the server were never cleared,
and subsequent requests served be the same interpreter could lead
to the old error states.

Many thanks to Oleg Oleinick for reporting and the great test cases.