Silence harmless entry in error log concerning unexpected EOF
When a client connects to NaviServer and terminates the connection
before the initial TLS/SSL handshake is finished, NaviServer wrote a
message complaining about UNEXPECTED_EOF_WHILE_READING. When a client
just probes NaviServer, this message showed up. In the new version,
this message is printed during handshake only in Debug mode.