Improved Tcl9 compatibility of ns_http regression test
Tcl 9 rejects to get byte-arrays via Tcl_GetByteArrayFromObj() when the
value of the Tcl_Objs has "invalid binary data". Although, one would
expect that every byte sequence is accepted as binary data, this has to
do with the fact, that the Tcl_Obj has already invalid UTF-8 that is
rejected by Tcl_GetByteArrayFromObj(). The solution is to make sure
that the data added to the Tcl_Obj in the first place is already
a proper byte array. This can be achieved e.g. via

    set binary_data [binary format a* $value]

In Tcl8, all data can be converted to a byte array with Tcl_GetByteArrayFromObj().
The variant with the explicit byte-array conversion works identically with
Tcl8 and Tcl9.