Fixed problem with parsing boundaries with binary data
The problem was introduced by comment b8ebf99 which provides a
performance improvement for parsing of multipart entries.  The
improvement had problems with some binary data. The new version is
using the emulation of the function memmem(), which is available on
some platforms, but not everywhere. A future update should test for
the availability of memmem() in the C-library and fall back to the
emulation when necessary.