Fixed potential race condition with "ns_cache_eval -force"
There was a problem with "ns_cache_eval -force", where the system
relied on the existence of a pre-existing entry, but where it was
flushed in the meantime by various potential means. Now the value
during the "-force" call is cleared exactly like in the case of an
unset operation. The null-value operations are already protected until
these are finished on various places in the code.

The regression test was extended to handle various such concurrency
situations. When the problem still exists, one should see a log
message like "old entry ...different from re-fetched entry ..."