Fixed ns_crypto::aead::encrypt/decrypt test under OpenSSL 1.1.1 on Ubuntu 18.04.4
In short, the problem was that with this version of OpenSSL, setting
empty additional authenticated data (AAD) behaved differently from
other versions, namely it was clearing incorrectly the information
that the initialization vector (IV) was already set. An upgrade of
OpenSSL fixed the problem. However, with these changes, also the stock
version of OpenSSL can be used.

As a byproduct, better error messages are now produced, the code
received more cleanup (e.g. explicit initialization, etc.)