ns_http: Fixed behavior of HEAD request for persistent connections.
The old code expected to read a content independent of the request
method, leading to a timeout if there was none, as it is the case on
HEAD requests. This was not a problem before keep-alive, since the
connections were always closed. The regression test was extended to
include tests for persistent connections.