Address not respected timeouts for HTTPS requests
As addressed by Marty in the openacs.org forums, a problem existed
when making ns_http requests of TLS. At least in the analyzed cases,
this could happen, when the TCP connection to a server succeeded
quickly, but the TLS handshake was taking a long time, without a
raising an error. The new code introduces a timeout during TLS
setup/handshake. This requires changes for Ns_TLS_SSLConnect() call to
pass into it an optional timeout value and change the result to be
able to report back a NS_TIMEOUT result value.