nscp: fix problem, when the nscp is activated but no "users" section is specified
In cases, where the nscp module is activated (by loading the module),
but the configuration file has no section for configuring, which users
should get access, nsd was crashing during startup.

This bug exists in the repository since 2022-07-11 in the main branch,
but was not included in any released version. Fortunately, the
problem could be fixed mostly by deleting code.

Many thanks to Andrew Piskorski for reporting the problem.