Improved handling of binary data with Tcl 8.7 or newer
This change makes the autodection of binary data more conservative
and more similar to Tcl 8.6. The auto-dectected byte arrays
must now have "bytes" set to NULL. This addresses a problem,
where a UTF-8 string might be converted to a byte-string
but where still the UtfToExternalDString() etc. have to be
called. In Tcl 8.7 or newer, the following command
converts as a side-effect the Tcl_Obj "string" into an byte-array, but the
semantics must be the same with or without this command

    set s1 [encoding convertfrom utf-8 $string]

This problem does not exist with Tcl 8.6. This change
introduces as well some more detailed debugging aids.