Added explicit handling of completely received headers
Before, handling of completely received headers was controlled by the
value of the content-length and received content. Since we have to
handle content lengths precisely for persistent connections, and there
are cases, where content-length is set to 0, or not provided all, or
not provided yet (e.g. long list of header fields, requiring multiple
read operations), we had to add explicit information to keep track of
this fact, when the headers are ready to be processed.