ns_http: added option "-partialresults"
This new option allows returning also partial results from a request
that might have been terminated in the middle. When specified, no
exception is raised, and the error is returned as the dict member
"error".  This option is useful to inspect partial results also when
e.g. the server terminates the connection unexpectedly. This option is
necessary for regression testing under windows, where a socket
connecting two threads leads immediately to an error condition on the
other end of the socket.