Provide compatibility between in-memory and file-based form-data handling
Since the file based implementation uses a fallback-charset of iso8859-1
when parsing form data in POST frequests (in order to be able to extract
"_charset_" data), the in-memory based variant does now the same. The
decision of which parser is used is taken based on the size of the
form data.... and should therefore be consistent.