Reduced usage of double-locked checking pattern for one-time initialization
This change introduces the macro NS_INIT_ONCE(fn).  This macro
implements one-time initialization in a thread-safe manner.  The macro
addresses the concerns expressed in the "Double-checked Locking"
pattern.

   https://en.wikipedia.org/wiki/Double-checked_locking

The provided function "fn" should return NS_TRUE for compatibility with
windows. For Unix compilations, the return value is ignored.

NS_INIT_ONCE is defined as a macro to provide different variables for
the controlling variables.

NS_INIT_ONCE is used for one-time initialization of InterpData.