New sub-command: ns_logctl grep ?-filename fn? /string/
The new sub-command reads the content of the system log file, joins
continuation lines, strips color codes and greps in the result for the
provided string. The function returns all lines containing the
specified string. When "-filename" is specified, the search is
performed on the specified file. The specified filename should be
fully qualified.

The new sub-command is used e.g. by the nsstats module.