Experimental feature: dynamic certificate adding for mass virtual hosting
So far, NaviServer supported mass virtual hosting (without restarts)
for nssock, but not for nsssl. With this change, it becomes possible to
add at runtime additional certificates for a driver, such that these
can be picked up at runtime as virtual servers via SNI. The
connections are mapped to the "defaultserver" of the driver.

To enable this feature, add the parameter "vhostcertificates" to the
driver module, having as value a directory containing .pem files of
virtual hosts. In the following snippet, this directory is named
"certificates".

    ns_section ns/module/nsssl {
        ns_param defaultserver	$server
        ...
        ns_param certificate	$serverroot/etc/certfile.pem
        ns_param vhostcertificates $serverroot/etc/certificates
        ...
    }

When a request for an unknown SNI server name comes in, NaviServer
performs now a lookup in the specified directory. If it finds a
certificate named $serverroot/etc/certificates/SERVERNAME.pem, this
server name is accepted as a virtual host and is mapped to the default
server of this driver.

TODO:
- update documentation
- list expiration dates in nsstats