Fix for fastpath in connection with "ns_serverrootproc"
Earlier versions of NaviServer initialized for fastpath the
interpreter with the connection conditionally, and rather late.  When
"ns_serverrootproc" is configured, the interpreter with its linkage to
the connection is needed very early in the request.  Now the
initialization for fastpath is made early, when this callback or the
callback for the connection location proc is set.

Many thanks to Georg Lehner for reporting this bug and testing.