Added auto-SNI support to "ns_connchan open"
Previously, auto-SNI was added to ns_http (to use the non-numeric
hostname as default for SNI). Now the same feature is used in
"ns_connchan open" as well. The change required some refactoring. This
change eases the use of "ns_connchan" with cloud services, and affects
all packages using it (e.g. reverse proxy).