ns_conn details: include servername in result dict for HTTPS
Background: the "servername" entry provides the name as it was used in
the SNI (server name indication) connections setup of an https
connection. Compared to the host header field, the value from SNI is
more trusted, since e.g., for mass hosting, the SNI value is used to
determine the certificate.