ns_conn protocol: Fix documentation and backwards compatibility bug
In very old versions of NaviServer, "ns_conn protocol" returned just
in the proxy server case the specified protocol. In 2020, it was
changed to return always the current protocol, without updating the
documentation. Now, in proxy server cases it behaves like before, in
other cases it returns the actual protocol.