ns_conn reform for non-proxy requests
So far, NaviServer has provided useful values "ns_conn
host|port|protocol" only for proxy requests, i.e., when NaviServer is
running as a proxy-server (which is barely used today - don't mix up
with reverse-proxy, or behind a reverse proxy server). With these
changes, nothing changes for the proxy requests, but for non-proxy
requests, these values are now returning the cooked info from the
"host" header field (parsing of the header fields, stripping trailing
dots, handling IP-literal notation). Notice that it is still possible
to use the method used commonly today by processing the host header
field on the Tcl level.

These changes will be only in NaviServer 5

- ns_conn host ?default?: The "ns_conn host" command receives a
  default value, in case, not host header is provided

- ns_conn target: new sub-command to return the request target as
  specified in [1] (i.e., return the request URL optionally followed
  by query variables)

As an example, a redirecting server from HTTP to HTTPS can be now
implemented in a few lines:

=====================================================================
  ns_section ns/servers {
     ns_param    redirector      "Redirect to HTTPS"
  }
  ns_section ns/server/redirector/modules {
     ns_param    nssock          nssock
  }
  ns_section ns/server/redirector/module/nssock {
     ns_param    address         0.0.0.0
     ns_param    port            8000
  }
  ns_section ns/server/redirector/tcl {
      ns_param initcmds {
          ns_register_proc GET /* {
              set host [ns_conn host localhost]
              ns_returnmoved https://$host:8443[ns_conn target]
          }
      }
  }
=====================================================================
[1] https://datatracker.ietf.org/doc/html/rfc7230#section-3.1.1