nscgi: fixed handling of exit-codes =! 0, and invalid CGI content
So far, the exit codes of CGI scripts were ignored. Now, the nonzero
exit codes are handled unless the output is a valid CGI
output. Reasoning: When the result is a valid CGI output, the CGI
output might set the HTTP status, and will not return a different
status code due to the nonzero exit of the script.

CGI Content is treated as invalid, when it has no plausible headers.

The regression test was extended accordingly