Bug fix: Ns_SetIUpdateSz() with an unknown keyLength could lead to a crash
This bug happens, when a user sets the content-type header field with
a lowercase field name, and when later the same field is updated with
a capitalized value, and this happens via the (old) C API, which does
not provide a length of the key of the set.

Many thanks to Brian Fenton for the help (report and patience)!