nscgi: new command: ns_register_cgi ?-noinherit? ?-path path? ?--? method url"
The new command is used to register the CGI request handler in analogy
to e.g. ns_register_fastpath. This allows us
- to map/unmap the CGI request handler at runtime (unmap via ns_unregister_op),
- to test dynamically more scenarios in the regression test suite,
- to use the usual inherit/noinherit options as for the other register_* operations.

The command uses the nscgi module registered for the server,
in which the command is executed.

Limitations:
the "ns_register_cgi" command supports currently only one nscgi module
per server. When the module is loaded globally (i.e., without an associated
server), the command is not registered.

TODO: the description of the command is not yet included
in the documentation.