New option "-all" for the command "ns_server ... mapped ... {METHOD PATH}"
This change improves the introspection abilities for the command
"ns_server ... mapped ...". When the option "-all" is specified, it
returns a dict containing the connection pool and the request handler
responsible for handling the specified METHOD and PATH.

Minor refactoring was necessary to obtain the value of the mapping of
the request handles as with the pool mapping.

When the option "-all" is not specified, the command is fully
compatible with earlier versions.

Extended regression test for the new option.