Increased reusability: factored out function Ns_RegisterFastUrl2File()
The new function provides functionality on the C API, which was
previously hard-wired to the Tcl API implementation. With this
function, one can register an Ns_RegisterUrl2FileProc() with either
Ns_FastUrl2FileProc() or NsMountUrl2FileProc() depending on the
provided basePath.  Before, this logic was part of
NsTclRegisterFastUrl2FileObjCmd, the implementation of the Tcl API
"ns_register_fasturl2file"