New functionality: path-segment match for urlspace
Background: Before this change, all urlspace matches were performed at
leaf level. In other words, all wildcard matches were only applied at
the last segment of the URL, pointing typically to a filename. These
tail matches are performed no matter if "inherit" was specified or
not. Examples:

   registered pattern: /somedir/*.tcl
   matches:            /somedir/foo.tcl
                       /somedir/sub/bar.tcl
   no match            /somedir/baz.tcl/some-tail

Interestingly, when no wild cards were specified, the behavior was different

   registered pattern: /somedir/test.tcl
   matches:            /somedir/test.tcl
                       /somedir/test.tcl/some-tail

The reason for this is that urlspace always performs a wildcard match,
and the first pattern corresponds to /somedir/test.tcl/*

For CGI, it is necessary to perform not only tail matches, but segment
matches as well. For segment-matches, the part after the matching
pattern has to be passed according to RFC 3875 in the PATH_INFO
variable. So far, the nscgi package was performing its own matching
logic, leading to several inconsistencies.

With the new implementation, NsUrlSpecificSet() can set the new flag
"NS_OP_SEGMENT_MATCH", which causes NsUrlSpecificGet() to perform a
match on all segments (provided it is not a fully blank pattern, which
would match everything). This covers e.g. the "no match" case of
above.