Added templating for ns_returnnotice
So far, NaviServer used a hard-coded template for the HTML of
ns_returnnotice, which is used e.g. for error messages or for
directory listings. Now, this template is realized as .adp file, which
resides per default in SERVERHOME/conf/returnnotice.adp. This template
file can be configured via the parameter "noticeadp" in the
configuration section of the server.

  ns_section ns/server/$server {
    # ...
    ns_param noticeadp returnnotice.adp ;# ADP file for ns_returnnotice.
    # ...
  }

When the configured name is not absolute, it is tried to be looked up
from SERVERHOME/conf/. When the parameter is set to an empty string,
the old-style hard-coded template is used.