Apply rejectalreadyclosedconn also on detached connections
Upon request on the naviserver-devel mailing list, the backwards compatibility
configuration flag rejectalreadyclosedconn is now applied also on "detached"
connections.  Previously, it was only applied on closed connections.

As explained in

    https://openacs.org/forums/message-view?message_id=5659253

it is not recommended to set this flag, but to fix the script code to
avoid writing to connections which are not available to this thread anymore.
The right thing is to add a return (in an executing script) or an ad_script_abort
in library functions (e.g. in OpenACS) or similar after e.g. a "ns_return*"
statement.