Fix for request parsing proxy requests and CONNECT requests
The updated version fixes a potential assertion violation in the main
branch and generalizes request parsing to cover also the syntax of
CONNECT requests, where the provided connect target + port are
returned as members of the parsed request structure. The assertion was
added as a sanity check to test more tighten, whether some assumptions
hold.

The type of the request (plain request type, proxy request, and
connect request) is returned as well in this structure to avoid a
duplication of the logic.