Add support for adding an entry to the access.log from outside connection threads
The new function AddNslogEntry() can be used to add an entry to the
access log, when the request is not handled by the trace of a
connection thread. The function respects the settings of the nslog
module for the server and behaves gracefully, when nslog module was
not loaded or not registered for the server causing the log entry.

Applications:
  - direct replies from the driver thread (via SockSendResponse())
  - 100 CONTINUE,
  - cases where the TRACE is not called in the connection thread.

Currently, the function is just used for the first two cases.

New functions to support this functionality
- AddNslogEntry()

- NsGetProcFunction():
  Returns the function pointer of a function with the specified description value.

- NsGetTraceProcArg():
  Returns the "arg" value which was registered for the specified function pointer.