NsAddNslogEntry(): Generalize handling and reduce exposing of function pointers
The reworked implementation is based on an extension of the classical
NsRunTraces() by providing an additional argument to call just certain
traces. This is implemented by the new API function
NsRunSelectedTraces(). This takes care of multiple registrations of
the same function for a server. Previously, just the first trace
callback of the specified type was called.

The previously introduced API call NsGetTraceProcArg() is not needed
anymore and was removed.