Refactor code for persistent connections
The previous code had problems, when concurrent threads started/reused
persistent connections to the same host, especially in error
conditions (e.g. when the external host providing a web API was
overloaded). The refactored code is much more robust by keeping all
updates to the keep-alive socket handing under the same
mutex. Furthermore, the code is significantly simpler by not using an
additional Ns_Task for closing timed-out keep-alives. Currently, the
keep-alives are just precise on the second granularity, but that might
be changed in the future.

The regression test was extended for covering some more potential
error cases.