nsproxy: added subcommand "ns_proxy workers /pool/"
The new subcommand returns a dict per worker contain e.g. "pid" and
usage statistics such as e.g.

    id ExecPool-0 pid 0 created 0 runs 0 state idle
    id ExecPool-1 pid 0 created 0 runs 0 state idle
    id ExecPool-2 pid 0 created 0 runs 0 state idle
    id ExecPool-3 pid 0 created 0 runs 0 state idle
    id ExecPool-4 pid 0 created 0 runs 0 state idle
    id ExecPool-5 pid 86249 created 1 runs 1 state idle
    id ExecPool-7 pid 84659 created 1 runs 84 state idle

This can be used for fine-tuning the required number of workers for an
application by checking the work load.

Support for nsstats to contain these statistics when available were also added.