Rework "rejectalreadyclosedconn"
The previous implementation was just about skipping raising exceptions
on attempts of writing to a client when the connection is not
available anymore. Skipping the exception has the problem, that some
operation might run into problems, causing a crash. The new
implementation causes now commands attempting to write to a connection
after a command causing a connection close/detatch/ to become
effectively no-ops. It should be safe now (but not recommended) to set
"rejectalreadyclosedconn" to false. When this condition happens, such
commands write a single line into the system log.

The recommended action is to keep the default of
"rejectalreadyclosedconn" and to fix the (probably) legacy code,
avoiding such impossible output operations.