Don't let CloseWaitingCheckExpire() close something, which as marked as "INUSE"
The janitor taks cleaned "close-waiting" entries marked as free to cover
cases with stale INUSE cases. However, at least OpenSSL 1.1.1 can
crash in some cases, when operating on already closed connections.
With the change there is the danger that broken connection might
hang around for a long time, but we will see if this can be a problem.