let shutdown callbacks use the spoolers
When using a shutdown callback from Tcl (e.g. ns_atshutdown), the
i/o operations might still required the spooling infrastructure.
In essence, this change moves the shutdown of the spooling
infrastructure to a later stage (after NsRunAtExitProcs).
This change was necessary for the cluster mode in OpenACS
when using dynamic cluster nodes.