Improved portability: don't assume that type "char" is the same as "unsigned char"
On most platforms, "char" is the same as "unsigned char". However, the
standard does not enforce it, since it defines three distinct types:
char, signed char, and unsigned char. The first platform that I found,
where this is different, is aarch64 (ARM) with musl (Alpine
Linux). With x86_64 and musl "char" is signed.