Added a new configure macro NS_MSG_IOVLEN_T
The previous code assumed that the size of the member "msg_iovlen" of
"msghdr" is size_t, unless the following expression holds:

    defined(__APPLE__) || defined(__darwin__)
    || defined(__FreeBSD__) || defined(__NetBSD__) || defined(__OpenBSD__)

However, at least with Alpine Linux (musl) the type is "int". The new
macro NS_MSG_IOVLEN_T is 1, when "msg_iovlen" is of type "size_t".  So
far, the new m4 macro NS_MSG_IOVLEN_T is used only for the default
case.