"x-forwarded-for" reform (part 1)
The x-forwarded-for header is used by various servers to identify the
originating IP address of a client connecting to a web server through
an HTTP proxy or a load balancer. The usage of the header grew lately
with CDNs, Cloud Services (AWS, Azure, and Google Cloud), or Web
Application Firewalls becoming more popular for an extending range of
services. Typically, a web server tries to determine the client
address for security, logging, debugging and geolocation purposes, and
helping to understand the usage patterns of web servers.

Traditionally, NaviServer has implemented x-forwarded-for handling in
the simplest way, by treating the first (leftmost) value in the
x-forwarded-for header as the client (see also [1] for a brief
description). The old approach proved to work well in many production
environments.

There are two potential problems with this:

a) If the client is connecting from a private network through
   potentially multiple proxy servers, the private address
   (e.g. 192.168.1.100) may end up as the peer (and also e.g. in the
   access.log), which is not useful for any kind of analysis.

b) If the client itself sends a (faked) "X-Forwarded-For" header, it
   can provide whatever it wants as header value.

One measure is for a reverse proxy to only accept X-Forwarded-For
headers from trusted sources, or to always treat its peer as a client
(providing the peer address in the header field).

This can lead to some loss of information, since the real client is
not determined if there are multiple proxy servers in the chain.

Changing the configuration of an upstream proxy server it is not always
possible, as the list of applications above hints.

The new version of NaviServer has now more options to obtain the
client address of the forwarded-for header. In general, we have no a
left-to-right mode and a right-to-left mode.

- The default mode processes the address list from left to right as
  before (so, this is full backward compatible). Additionally, one can
  specify to skip non-public IP addresses which are not useful for
  geolocation purposes (see point (a) above).  If the received
  "x-forwarded-for header contains, for example, "192.168.1.100,
  1.1.1.1, 2.2.2.2", then "1.1.1.1" will be used as the client IP
  address.

- The right-to-left mode is selected, when the NaviServer site
  administrator configures trusted reverse proxy servers in the
  configuration file either by IP address or in CDIR notation.

  As a consequence, the "x-forwarded-for" header will only be accepted
  from trusted servers, otherwise the header will be ignored. In
  addition, the address list is processed from right to left, skipping
  known trusted reverse proxies. This reduces the problem of spoofed
  addresses. In the address list "142.251.208.110, 1.1.1.1,
  192.168.2.10", the trusted reverse proxy "192.168.2.10" is skipped
  and "1.1.1.1" is used (instead of the spoofed Google IP on the
  left).

  Additionally, one can configure to ignore non-public IP addresses like
  in the left-to-right mode above. When all addresses are skipped, the
  behavior is the same, as if no x-forwarded-for header is present.

Note: Only when NaviServer receives multiple entries in the
"x-forwarded-for" header is there a difference between these two
modes.

In order to ease configuration and naming conventions, the old
parameter "reverseproxymode" was changed into a section
"ns/parameters/reverseproxymode" keeping the new parameterization
options. The old parameter can still be used for backward
compatibility.

OLD:
    ns_section ns/parameters {
       ...
       ns_param reverseproxymode true
       ...
    }

NEW:
    ns_section ns/parameters/reverseproxymode {
        ns_param enabled        on
        ns_param trustedservers {192.168.0.0/16 137.208.89.213}
        ns_param skipnonpublic  true
    }

Still open:
 - documentation updates
 - updated sample configuration files

[1] https://en.wikipedia.org/wiki/X-Forwarded-For